+++ changes.md
@@ -7,6 +7,14 @@
 - `xui-iconwrapper-medium` width and height updated to 32px (from 30px)
 - `xui-styledcheckboxradio-reversed` left padding updated to 15px (from 7px) when it's under `xui-styledcheckboxradio-group`
 - `xui-styledcheckboxradio--label` remove the left margin when reversed
+- `xui-styledcheckboxradio--input + xui-iconwrapper`
+  - For uncheked icon checkbox, remove the background color when hover or active
+  - For checked icon checkbox, modify the svg stroke color when it's focused and keep the box-shadow when it's hovered
+- `xui-styledcheckboxradio--radio`
+  - For unchecked and checked radio, remove the background color when it's hovered and active
+  - For checked radio, the background color now will change when it's hovered and active
+- `xui-switch--control`
+  - For unchecked and checked switch, the background color now will change when it's hovered and active
 
 ### Removals